Use C99 and -m64 for all platforms
I have temporarily removed the illumos-specific bits from mmap because
they rely on interfaces not currently exposed when enabling SUSv3
(_XOPEN_SOURCE=600) and __EXTENSIONS__. I think this is a bug because
those extensions do not interfere with SUSv3, but for now it's easiest
to simply disable this additional functionality. I need to revisit
this later and determine if it's worth keeping this additional
functionality around.