Remove time correlation stat
Fixes #31

This statistic uses a sum of least squares regression line to
correlate the passage of time (independent variable) with the latency
of the call (dependent variable). I can't be sure of their exact
reasons, but my guess is the original authors intended for this to
tell you if the benchmark got slower or faster over time. A
significant positive number might indicate a resource being leaked or
some other cause of slowdown, a significant negative number a cache
being warmed.

In practice I see this number fluctuate too much across the same
benchmark to be of any practical use. I'd rather stick to a few well
understood stats.